refactor(desktop-driver): split oracle.go along its real input/build seam
CodeScene's health kept moving weight across three passes on this single
file (validateOracleOptions -> buildManagedFileOracle -> pairManagedPaths)
without clearing the 8.00 bar, because CodeScene scores per file and the
complexity was spread across the whole file, not concentrated in one
function. More in-file extraction would keep moving it around.

Read the whole file to check for a real seam rather than assuming one.
Found it, and confirmed it mechanically: every function/type that touches
the agent/permission/claudecode/daemonaddr domain is exactly
{oracleBuild, buildManagedFileOracle, newOracleBuild, writeEveryApplySubset,
writeApplySubset, declaredManagedPaths, resolveWrittenPaths,
managedApplyPermissions, applyManagedPermissions} plus the functions that
call them — everything else (option validation, manifest-row parsing) has
zero dependency on that domain. Nothing in the file except
runManagedFileOracle is called from outside it, so this is a real internal
seam, not cross-package reuse pressure.

Split into two files along that seam:
- oracle_options.go: the oracle's validated inputs — CLI flags
  (managedFileOracleOptions, runManagedFileOracle, validateOracleOptions and
  its six single-check helpers) and the baseline manifest format
  (managedFileEntry, parseManagedFileRow, readManagedFileManifest). Zero
  imports from irrlicht/core.
- oracle.go: building and writing the oracle's expected state given those
  already-validated inputs — the Claude Code agent declaration, Apply-closure
  enumeration, shadow-HOME path pairing, baseline materialization, and state
  writing.

Each file stands on its own: the build file's shape (validate -> read
manifest -> resolve agent -> exercise every Apply subset -> copy the
all-applied state) doesn't require knowing how a flag is validated or a
manifest row is parsed; the options file's shape (flags in, entries out)
doesn't require knowing how the build uses them. Verified with a mechanical
diff of function/type declarations before and after: the same set exists,
none dropped or duplicated, none merged, weakened, or reordered — this is a
pure move, not a rewrite.

The path-escape sanitizer (pairManagedPaths' "outside real HOME" / "did not
relocate exactly" checks) and its sink (the pairs map written to shadowPath)
both stay in oracle.go, unseparated. The CLI bind-address/output-path
validations remain the unconditional first call in buildManagedFileOracle,
just as visible as before a reader crosses one file to see their bodies.

Verified: go build/vet clean, go test
./tools/onboarding-factory/cmd/desktop-driver/... -race -count=1 (all 7
cases) and the full ./tools/onboarding-factory/... suite are locks — they
pass by construction under this pure move, not as red-first proof.
